+++ index.ts
@@ -1,5 +1,4 @@
-import 'isomorphic-fetch'
-import * as qs from 'qs'
+import * as axios from 'axios'
 import { BingToGeoJSON, BingOptions } from './providers/bing'
 import { GoogleToGeoJSON, GoogleOptions } from './providers/google'
 import { MapboxToGeoJSON, MapboxOptions } from './providers/mapbox'
@@ -117,8 +116,8 @@ export async function bing(address: string, options = BingOptions): Promise<GeoJ
  * @returns {Promise<GeoJSON.FeatureCollection<GeoJSON.Point>>} GeoJSON Results
  */
 async function get(url: string, params: any, geojsonParser: any, options: any): Promise<GeoJSON.FeatureCollection<GeoJSON.Point>> {
-  const response = await fetch(`${ url }?${ qs.stringify(params) }`)
-  const json = await response.json()
+  const response = await axios.get(url, { params })
+  const json = response.data
   const geojson: GeoJSON.FeatureCollection<GeoJSON.Point> = geojsonParser(json, options)
   return geojson
 }